+++ src/main/java/ru/yandex/practicum/filmorate/storage/review/ReviewLikeStorage.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage.review;
 public interface ReviewLikeStorage {
 
-    void addLike(int reviewId, int userId, boolean isPositive);
+    void addLike(long reviewId, long userId, boolean isPositive);
 
-    void deleteLike(int reviewId, int userId);
+    void deleteLike(long reviewId, long userId);
 }
